amd/microcode: Change equivalant ID width to 16bit
The definition of processor_rev_id in struct microcode
is 16 bits. So we need to change the a series of parameters
passing to 16 bits.

Reviewed-on: https://review.coreboot.org/c/coreboot/+/41869
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/cpu/amd/pi/00730F01/update_microcode.c
@@ -37,7 +37,7 @@ static u16 get_equivalent_processor_rev_id(u32 orig_id)
 
 void update_microcode(u32 cpu_deviceid)
 {
-	u32 equivalent_processor_rev_id =
+	u16 equivalent_processor_rev_id =
 				get_equivalent_processor_rev_id(cpu_deviceid);
 	amd_update_microcode_from_cbfs(equivalent_processor_rev_id);
 }
